LayoutTests/imported/w3c/web-platform-tests/html-aam/fragile/area-role-expected.txt



FAIL el-area assert_equals: <area shape="rect" coords="0,0,15,15" href="#" alt="x" data-testname="el-area" data-expectedrole="link" class="ex"> expected "link" but got ""
PASS el-area-no-href
